recipeServerPython/requirements.txt
Django>=4.0
djangorestframework>=3.13
psycopg2-binary>=2.9 # For PostgreSQL connection
django-admin-log # Optional, for viewing logs in admin later
Pillow>=9.0 # For ImageField support 